// generated/src/app/api-models/cptdt-quality-rpt-wrk.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CptdtQualityRptWrk {
  seqCapCptdtQltyWrk: number;
  seqCapCptdtRptWrk: number;
  seqIncentiveRule: number;
  seqQualityPgm: number;
  seqProvContract: number;
  qualityIncentiveAmount: number;
  qualityIncentivePercent: number;
  qualityCapApply: string;
  incentiveRetro: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}